Wallet: fix card type parsing
The card type actually consists out of two bytes in big endian.
The XOR key is 0x3d3f. This is relevant because the employee type
actually exceeds one byte with 400 denoting an employee of the
university part of KIT and 403 an employee of the large-scale
research part (to allow for different billing).

200 is a regular guest, add it as such.

Signed-off-by: Philipp Kern <phil@philkern.de>